Renamed time_utils mod into timefmt
Closes #50.
